Reema Alhenaki/FastAPI_app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pandas
httpx
pytest